修复Windows复制模板失败问题
