+++ chat2db-server/chat2db-server-web/chat2db-server-web-api/src/main/java/ai/chat2db/server/web/api/controller/rdb/RdbDocController.java
@@ -15,10 +15,7 @@ import ai.chat2db.server.web.api.controller.rdb.factory.ExportServiceFactory;
 import ai.chat2db.server.web.api.controller.rdb.request.DataExportRequest;
 import ai.chat2db.server.web.api.controller.rdb.vo.TableVO;
 import ai.chat2db.spi.model.Table;
-import ai.chat2db.spi.sql.Chat2DBContext;
-import ai.chat2db.spi.util.JdbcUtils;
 import cn.hutool.core.date.DatePattern;
-import com.alibaba.druid.DbType;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
